+++ BEFALL23-DATN-SD05/src/main/java/com/example/befall23datnsd05/controller/KhuyenMaiController.java
@@ -6,6 +6,7 @@ import com.example.befall23datnsd05.enumeration.TrangThai;
 import com.example.befall23datnsd05.enumeration.TrangThaiKhuyenMai;
 import com.example.befall23datnsd05.service.ChiTietSanPhamService;
 import com.example.befall23datnsd05.service.KhuyenMaiService;
+import com.example.befall23datnsd05.worker.PrincipalKhachHang;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.query.Param;
@@ -36,18 +37,30 @@ public class KhuyenMaiController {
 
     List<TrangThaiKhuyenMai> list = new ArrayList<>(Arrays.asList(TrangThaiKhuyenMai.DANG_HOAT_DONG, TrangThaiKhuyenMai.DUNG_HOAT_DONG, TrangThaiKhuyenMai.SAP_DIEN_RA));
 
+    private PrincipalKhachHang principalKhachHang = new PrincipalKhachHang();
+
     @GetMapping("")
     public String hienThi(Model model) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         model.addAttribute("listKhuyenMai", service.getList());
         model.addAttribute("listTrangThai", list);
+        model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
         return "admin-template/khuyen_mai/khuyen_mai";
     }
 
     @GetMapping("/trang-thai/{trangThai}")
     public String getByTrangThai(Model model,
                                  @PathVariable("trangThai") TrangThaiKhuyenMai trangThaiKhuyenMai) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         model.addAttribute("listTrangThai", list);
         model.addAttribute("listKhuyenMai", service.getByTrangThai(trangThaiKhuyenMai));
+        model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
         return "admin-template/khuyen_mai/khuyen_mai";
     }
 
@@ -56,6 +69,10 @@ public class KhuyenMaiController {
                              @Param("trangThai") TrangThaiKhuyenMai trangThaiKhuyenMai,
                              @Param("startDate") LocalDate startDate,
                              @Param("endDate") LocalDate endDate) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         if (startDate.isAfter(endDate)) {
             model.addAttribute("listTrangThai", list);
             model.addAttribute("startDate", startDate);
@@ -66,6 +83,7 @@ public class KhuyenMaiController {
         model.addAttribute("startDate", startDate);
         model.addAttribute("endDate", endDate);
         model.addAttribute("listKhuyenMai", service.findKhuyenMai(startDate, endDate, trangThaiKhuyenMai));
+        model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
         return "admin-template/khuyen_mai/khuyen_mai";
     }
 
@@ -74,24 +92,31 @@ public class KhuyenMaiController {
             @ModelAttribute("khuyenMai") KhuyenMaiRequest khuyenMaiRequest,
             Model model
     ) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         model.addAttribute("khuyenMai", new KhuyenMai());
+        model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
         return "admin-template/khuyen_mai/them_khuyen_mai";
     }
 
 
     @PostMapping("/add")
-    public String them(
-            @Valid
-            @ModelAttribute("khuyenMai") KhuyenMaiRequest khuyenMaiRequest,
-            BindingResult bindingResult,
-            Model model
-    ) {
+    public String them(@Valid @ModelAttribute("khuyenMai") KhuyenMaiRequest khuyenMaiRequest,
+            BindingResult bindingResult, Model model) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         String ten = khuyenMaiRequest.getTen();
         if (bindingResult.hasErrors()) {
+            model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
             return "admin-template/khuyen_mai/them_khuyen_mai";
         } else {
             if (service.existsByTen(ten)) {
                 model.addAttribute("errorTen", "Tên  đã tồn tại");
+                model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
                 return "admin-template/khuyen_mai/them_khuyen_mai";
             }
             service.add(khuyenMaiRequest);
@@ -100,11 +125,13 @@ public class KhuyenMaiController {
     }
 
     @GetMapping("/view-update/{id}")
-    public String viewUpdate(
-            @PathVariable("id") Long id,
-            Model model
-    ) {
+    public String viewUpdate(@PathVariable("id") Long id, Model model) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         model.addAttribute("khuyenMai", service.getById(id));
+        model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
         return "admin-template/khuyen_mai/sua_khuyen_mai";
     }
 
@@ -112,14 +139,20 @@ public class KhuyenMaiController {
     public String update(@Valid @ModelAttribute("khuyenMai") KhuyenMaiRequest khuyenMaiRequest,
                          BindingResult bindingResult,
                          Model model) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         String ten = khuyenMaiRequest.getTen();
         Long id = khuyenMaiRequest.getId();
         if (bindingResult.hasFieldErrors("ten") || bindingResult.hasFieldErrors("mucGiamGia")
                 || bindingResult.hasFieldErrors("ngayKetThuc")) {
+            model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
             return "admin-template/khuyen_mai/sua_khuyen_mai";
         } else {
             if (service.existsByTenAndIdNot(ten, id)) {
                 model.addAttribute("errorTen", "Tên  đã tồn tại");
+                model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
                 return "admin-template/khuyen_mai/sua_khuyen_mai";
             }
             service.update(khuyenMaiRequest);
@@ -130,6 +163,10 @@ public class KhuyenMaiController {
 
     @GetMapping("/huy/{id}")
     public String huyKhuyenMai(@PathVariable("id") Long id) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         service.huy(id);
         ctspService.autoUpdateGia();
         return "redirect:/admin/khuyen-mai";
@@ -138,15 +175,24 @@ public class KhuyenMaiController {
     @GetMapping("/them-san-pham-khuyen-mai/{idKM}")
     public String sanPhamKhuyenMai(Model model,
                                    @PathVariable("idKM") Long idKM) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         model.addAttribute("ctspKhuyenMai", ctspService.getAllSanPhamKhuyenMai(idKM));
         model.addAttribute("ctspCoKhuyenMai", ctspService.getSpCoKhuyenMai(idKM));
         model.addAttribute("khuyenMai", service.getById(idKM));
+        model.addAttribute("tenNhanVien", principalKhachHang.getCurrentNhanVienTen());
         return "admin-template/khuyen_mai/san_pham_khuyen_mai";
     }
 
     @GetMapping("/them-san-pham-khuyen-mai/them/{idKM}/{idCtsp}")
     public String updateIdKhuyenMai(@PathVariable("idKM") Long idKM,
                                     @PathVariable("idCtsp") Long idCtsp) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         ctspService.updateIdKhuyenMai(idKM, idCtsp);
         ctspService.updateGiaBan(idCtsp);
         return "redirect:/admin/khuyen-mai/them-san-pham-khuyen-mai/{idKM}";
@@ -154,6 +200,10 @@ public class KhuyenMaiController {
 
     @GetMapping("/them-san-pham-khuyen-mai/xoa/{idKM}/{idCtsp}")
     public String deleteIdKhuyenMai(@PathVariable("idCtsp") Long idCtsp) {
+        Long idNhanVien = principalKhachHang.getCurrentNhanVienId();
+        if (idNhanVien == null){
+            return "redirect:/login";
+        }
         ctspService.deleteIdKhuyenMai(idCtsp);
         ctspService.updateGiaBan(idCtsp);
         return "redirect:/admin/khuyen-mai/them-san-pham-khuyen-mai/{idKM}";
